save and load

+++ src/game/functions/gameFunctions.ts
@@ -1,24 +1,34 @@
 import { useUiTempStore } from '../../ui/state/uiTempStore'
-import { GameState } from '../GameState'
 import { GetInitialGameState } from '../InitialGameState'
 import { loadData } from '../loadData'
 import { WorkerMessage } from '../loadWorkerTypes'
-import { setState } from '../setState'
 import { useGameStore } from '../state'
-import { MAX_LOAD } from '../const'
+import { MAX_LOAD, TEST_DIF } from '../const'
+import { GameState } from '../GameState'
 import { regenerate } from './regenerate'
 import { advanceTimers } from './advanceTimers'
 
 // eslint-disable-next-line import/default
 import LoadWorker from './loadWorker?worker'
 
-let worker: Worker | undefined
+let worker = new LoadWorker()
+let tempState: GameState
+
+function createWorker() {
+    worker = new LoadWorker()
+    worker.onmessage = (e: MessageEvent<WorkerMessage>) => {
+        if (e.data.loadingData) useUiTempStore.setState({ loadingData: e.data.loadingData })
+        else if (e.data.state) start(e.data.state)
+        if (e.data.state) tempState = e.data.state
+    }
+}
+
+createWorker()
 
 function killWorker() {
-    if (worker) {
-        worker.terminate()
-        worker = undefined
-    }
+    if (worker) worker.terminate()
+
+    createWorker()
 }
 
 function start(state: GameState): void {
@@ -30,16 +40,11 @@ function start(state: GameState): void {
 }
 
 export const load = (data: object) => {
-    worker = new LoadWorker()
-    worker.onmessage = (e: MessageEvent<WorkerMessage>) => {
-        console.log(e.data)
-        if (e.data.state) start(e.data.state)
-        if (e.data.loadingData) useUiTempStore.setState({ loadingData: e.data.loadingData })
-    }
     const state = loadData(data)
     state.loading = true
-    const end = Math.min(state.now + MAX_LOAD, Date.now())
-    const loadingData = { loading: true, start: state.now, now: state.now, end, percent: 0 }
+    const now = state.now + TEST_DIF
+    const end = Math.min(now + MAX_LOAD, Date.now())
+    const loadingData = { loading: true, start: now, now: now, end, percent: 0 }
     useUiTempStore.setState({ loadingData })
     useGameStore.setState(structuredClone(state))
     worker.postMessage(state)
@@ -50,5 +55,5 @@ export const stopLoad = () => {
 }
 export const startAnyway = () => {
     killWorker()
-    setState((state) => start(state))
+    if (tempState) start(tempState)
 }
